scripts/requirements.txt
Pillow>=9.0.0
tqdm>=4.64.0
